+++ podaac/subsetter/subset.py
@@ -23,6 +23,7 @@ import functools
 import json
 import operator
 import os
+from itertools import zip_longest
 from typing import List, Tuple, Union
 import dateutil
 from dateutil import parser
@@ -730,7 +731,48 @@ def build_temporal_cond(min_time: str, max_time: str, dataset: xr.Dataset, time_
     return temporal_cond
 
 
-def subset_with_bbox(dataset: xr.Dataset,
+def get_base_group_names(lats):  # pylint: disable=too-many-branches
+    """Latitude groups may be at different depths. This function gets the level
+    number that makes each latitude group unique from the other latitude names"""
+    unique_groups = []
+    group_list = [lat.strip(GROUP_DELIM).split(GROUP_DELIM) for lat in lats]
+
+    # make all lists of group levels the same length
+    group_list = list(zip(*zip_longest(*group_list, fillvalue='')))
+    print(f"group_list == {group_list}")
+
+    # put the groups in the same levels in the same list
+    group_list_transpose = np.array(group_list).T.tolist()
+
+    diff_count = ['' for i in range(len(group_list))]
+    group_count = 0
+    # loop through each group level
+    for my_list in group_list_transpose:
+        for i in range(len(my_list)):  # pylint: disable=consider-using-enumerate
+            count = 0
+            for j in range(len(my_list)):  # pylint: disable=consider-using-enumerate
+                # go through each lat name and compare the level names
+                if my_list[i] == my_list[j] and not isinstance(diff_count[j], int):
+                    count += 1
+            # if the lat names is equivalent to only itself then insert the level number
+            if count == 1:
+                if isinstance(diff_count[i], int):
+                    continue
+                if 'lat' in my_list[i]:  # if we get to the end of the list, go to the previous level
+                    diff_count[i] = group_count - 1
+                    continue
+
+                diff_count[i] = group_count
+
+        group_count += 1
+
+    # go back and re-put together the unique groups
+    for lat in enumerate(lats):
+        unique_groups.append(f'{GROUP_DELIM}{GROUP_DELIM.join(lat[1].strip(GROUP_DELIM).split(GROUP_DELIM)[:(diff_count[lat[0]]+1)])}')
+    return unique_groups, diff_count
+
+
+def subset_with_bbox(dataset: xr.Dataset,  # pylint: disable=too-many-branches
                      lat_var_names: list,
                      lon_var_names: list,
                      time_var_names: list,
@@ -775,28 +817,54 @@ def subset_with_bbox(dataset: xr.Dataset,
     if lon_bounds[0] > lon_bounds[1]:
         oper = operator.or_
 
-    lat_var_prefix = [f'{GROUP_DELIM}{GROUP_DELIM.join(x.strip(GROUP_DELIM).split(GROUP_DELIM)[:-1])}' for x in
-                      lat_var_names]
+    # get unique group names for latitude coordinates
+    diff_count = [-1]
+    if len(lat_var_names) > 1:
+        unique_groups, diff_count = get_base_group_names(lat_var_names)
+    else:
+        unique_groups = [f'{GROUP_DELIM}{GROUP_DELIM.join(x.strip(GROUP_DELIM).split(GROUP_DELIM)[:-1])}' for x in lat_var_names]
+
     datasets = []
-    for lat_var_name, lon_var_name, time_var_name in zip(
-            lat_var_names, lon_var_names, time_var_names
+    total_list = []  # don't include repeated variables
+    for lat_var_name, lon_var_name, time_var_name, diffs in zip(  # pylint: disable=too-many-nested-blocks
+            lat_var_names, lon_var_names, time_var_names, diff_count
     ):
         if GROUP_DELIM in lat_var_name:
-            var_prefix = GROUP_DELIM.join(lat_var_name.strip(GROUP_DELIM).split(GROUP_DELIM)[:-1])
+            lat_var_prefix = GROUP_DELIM.join(lat_var_name.strip(GROUP_DELIM).split(GROUP_DELIM)[:(diffs+1)])
             group_vars = [
                 var for var in dataset.data_vars.keys()
-                if var.startswith(f'{GROUP_DELIM}{var_prefix}')
+                if GROUP_DELIM.join(var.strip(GROUP_DELIM).split(GROUP_DELIM)[:(diffs+1)]) == lat_var_prefix
             ]
+            total_list.extend(group_vars)
+
+            # include variables that aren't in a latitude group
             if variables:
                 group_vars.extend([
                     var for var in dataset.data_vars.keys()
-                    if var in variables and var not in group_vars and not var.startswith(tuple(lat_var_prefix))
+                    if var in variables and var not in group_vars and var not in total_list and not var.startswith(tuple(unique_groups))
                 ])
             else:
                 group_vars.extend([
                     var for var in dataset.data_vars.keys()
-                    if var not in group_vars and not var.startswith(tuple(lat_var_prefix))
+                    if var not in group_vars and var not in total_list and not var.startswith(tuple(unique_groups))
                     ])
+                total_list.extend(group_vars)
+
+            # group dimensions do not get carried over if unused by data variables (MLS nTotalTimes var)
+            # get all dimensions from data variables
+            dim_list = []
+            for var in group_vars:
+                dim_list.extend(list(list(dataset[var].dims)))
+
+            # get all group dimensions
+            group_dims = [
+                dim for dim in list(dataset.coords.keys())
+                if GROUP_DELIM.join(dim.strip(GROUP_DELIM).split(GROUP_DELIM)[:(diffs+1)]) == lat_var_prefix
+            ]
+
+            # include any group dimensions that aren't accounted for in variable dimensions
+            var_included = list(set(group_dims) - set(dim_list))
+            group_vars.extend(var_included)
 
         else:
             group_vars = list(dataset.keys())