+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [issue/153](https://github.com/podaac/l2ss-py/issues/153): Remove the asc_node_tai93 variable when blank in the SNDR collections for xarray.decode_times to decode
 - PODAAC-5538: Reduce memory footprint of l2ss by loading each variable individually to write to memory
+- [issue/155](https://github.com/podaac/l2ss-py/issues/155): lat var name prefix now generalized to unique group names. Group variables determined in 
+  subset_bbox using the unique group name.
 - PODAAC-5537: Fixed AQUARIUS_L2_SSS_V5 dataset
 ### Security